# Improve `importRecursive` performance with a regexp