--- a/pkg/jsonnet/imports_test.go
+++ b/pkg/jsonnet/imports_test.go
@@ -8,6 +8,20 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func BenchmarkTransitiveImports(b *testing.B) {
+	imports, err := TransitiveImports("testdata/importTree")
+	require.NoError(b, err)
+	assert.Equal(b, []string{
+		"main.jsonnet",
+		"trees.jsonnet",
+		"trees/apple.jsonnet",
+		"trees/cherry.jsonnet",
+		"trees/generic.libsonnet",
+		"trees/leaf.libsonnet",
+		"trees/peach.jsonnet",
+	}, imports)
+}
+
 // TestTransitiveImports checks that TransitiveImports is able to report all
 // recursive imports of a file
 func TestTransitiveImports(t *testing.T) {
@@ -20,6 +34,7 @@ func TestTransitiveImports(t *testing.T) {
 		"trees/apple.jsonnet",
 		"trees/cherry.jsonnet",
 		"trees/generic.libsonnet",
+		"trees/leaf.libsonnet",
 		"trees/peach.jsonnet",
 	}, imports)
 }