Instead of parsing the whole AST trying to find imports, find them with a regexp. This is significantly faster for very nested projects 
I added a benchmark in the tests. Here's the result:
```
$ benchstat bench1.txt bench2.txt
name                 old time/op  new time/op  delta
TransitiveImports-8  0.00ns ±47%  0.00ns ±30%  -25.85%  (p=0.000 n=92+90)
```

In real-life cases, I have gotten 10x faster run time (2.5s to 200ms) on some of our heavier projects (which are the ideal cases for caching, where this function is used)