@julienduchesne 2022-09-13
Converting to a draft. Will work on a different approach