@julienduchesne 2022-09-19
Implemented in https://github.com/grafana/tanka/pull/759